Rename mod to cookie-consent-service
It fixes a duplicated module name in @persistence.
